requirements.txt
PyYAML>=3.11
bundletester
jujuclient>=0.53.0
Jinja2
MarkupSafe
mock
flake8
nose
beautifulsoup4
